+++ vizro-core/examples/scratch_dev/app.py
@@ -1,66 +1,46 @@
-# Vizro is an open-source toolkit for creating modular data visualization applications.
-# check out https://github.com/mckinsey/vizro for more info about Vizro
-# and checkout https://vizro.readthedocs.io/en/stable/ for documentation.
+# # Vizro is an open-source toolkit for creating modular data visualization applications.
+# # check out https://github.com/mckinsey/vizro for more info about Vizro
+# # and checkout https://vizro.readthedocs.io/en/stable/ for documentation.
 
+import vizro.models as vm
 import vizro.plotly.express as px
 from vizro import Vizro
-import vizro.models as vm
+from vizro.tables import dash_ag_grid, dash_data_table
 
-df = px.data.iris()
+tips = px.data.tips()
 
-page = vm.Page(
-    title="Page with lots of extra information",
+page1 = vm.Page(
+    title="LAYOUT_FLEX_WRAP_AND_AG_GRID",
+    layout=vm.Flex(wrap=True),
     components=[
-        vm.Graph(id="scatter_chart", figure=px.scatter(df, x="sepal_length", y="petal_width", color="species")),
-    ],
-    controls=[
-        vm.Filter(
-            column="species",
-            selector=vm.Dropdown(
-                description="""
-                    Select which species of iris you like.
-
-                    [Click here](www.google.com) to learn more about flowers.""",
-                # You could also do this with vm.Tooltip(text=...)
-            ),
-        ),
-        vm.Filter(
-            column="species",
-            selector=vm.RadioItems(
-                description="""
-                    Select which species of iris you like.
-
-                    [Click here](www.google.com) to learn more about flowers.""",
-                # You could also do this with vm.Tooltip(text=...)
-            ),
-        ),
-        vm.Filter(
-            column="sepal_length",
-            selector=vm.RangeSlider(
-                description="""
-                    Select which species of iris you like.
-
-                    [Click here](www.google.com) to learn more about flowers.""",
-                # You could also do this with vm.Tooltip(text=...)
-            ),
-        ),
-        vm.Filter(
-            column="species",
-            selector=vm.Checklist(
-                description="""
-                    Select which species of iris you like.
-
-                    [Click here](www.google.com) to learn more about flowers.""",
-                # You could also do this with vm.Tooltip(text=...)
-            ),
-        ),
+        vm.AgGrid(id=f"outer_id_{i}", figure=dash_ag_grid(tips, id=f"inner_id_{i}", style={"width": 1000}))
+        # vm.AgGrid(id=f"outer_id_{i}", figure=dash_ag_grid(tips, id=f"qwer", style={"width": 1000}))
+        for i in range(3)
     ],
 )
 
-dashboard = vm.Dashboard(
-    pages=[page],
-    title="blah blah blah",
+page2 = vm.Page(
+    title="LAYOUT_FLEX_GAP_AND_TABLE",
+    layout=vm.Flex(gap="40px"),
+    # components=[vm.Table(figure=dash_data_table(tips, style_table={"width": "1000px"})) for i in range(3)],
+    components=[
+        vm.Table(figure=dash_data_table(tips, id=f"qwert_{i}", style_table={"width": "1000px"})) for i in range(3)
+    ],
 )
+
+page3 = vm.Page(
+    title="cross_id",
+    layout=vm.Flex(gap="40px"),
+    components=[
+        vm.Table(figure=dash_data_table(tips, id="qwert", style_table={"width": "1000px"})),
+        vm.AgGrid(figure=dash_ag_grid(tips, id="qwert", style={"width": 1000})),
+    ],
+)
+
+dashboard = vm.Dashboard(pages=[page1, page2, page3])
+
 
 if __name__ == "__main__":
     Vizro().build(dashboard).run()
+
+## TO CHECK: also for figure and/or Graph?