+++ vizro-core/src/vizro/models/types.py
@@ -13,7 +13,7 @@ from typing import Annotated, Any, Literal, NewType, Optional, Protocol, Union, 
 
 import plotly.io as pio
 import pydantic_core as cs
-from pydantic import Discriminator, Field, StrictBool, Tag, ValidationInfo
+from pydantic import Discriminator, Field, StrictBool, StringConstraints, Tag, ValidationInfo
 from pydantic.json_schema import SkipJsonSchema
 from typing_extensions import TypeAlias, TypedDict
 
@@ -629,6 +629,14 @@ ActionType = Annotated[
 # Extra type groups used for mypy casting
 FigureWithFilterInteractionType = Union["Graph", "Table", "AgGrid"]
 FigureType = Union["Graph", "Table", "AgGrid", "Figure"]
+
+# TODO-AV2 D 3: think about how to make this public and use in our inbuilt actions. Compare to _IdProperty.
+# Consider how it works if you just specify model name and not a dot separated string. Do we need a new composite type
+# for Union[list[_DotSeparatedStr], dict[str, _DotSeparatedStr]] too? Consider parameter target form too and whether
+# it should work similarly.
+_DotSeparatedStr = Annotated[str, StringConstraints(pattern="^[^.]+[.][^.]+$")]
+"""A string that must contain exactly one dot ('.'), with at least one character on both sides.
+For example: 'model-id.children'."""
 
 
 # TODO-AV2 A 1: improve this structure. See https://github.com/mckinsey/vizro/pull/880.