+++ vizro-core/tests/unit/vizro/models/_action/test_action.py
@@ -84,6 +84,9 @@ class TestLegacyActionInputs:
         [
             [""],
             ["component"],
+            ["component."],
+            [".property"],
+            ["component..property"],
             ["component_property"],
             ["component.property.property"],
         ],
@@ -97,6 +100,9 @@ class TestLegacyActionInputs:
         [
             "",
             "component",
+            "component.",
+            ".property",
+            "component..property",
             "component_property",
             "component.property.property",
         ],
@@ -148,8 +154,17 @@ class TestLegacyActionOutputs:
             ([], []),
             (["component.property"], Output("component", "property")),
             (
-                ["component.property", "component.property"],
-                [Output("component", "property"), Output("component", "property")],
+                ["component_1.property_1", "component_2.property_2"],
+                [Output("component_1", "property_1"), Output("component_2", "property_2")],
+            ),
+            ({}, {}),
+            (
+                {"output_1": "component.property"},
+                {"output_1": Output("component", "property")},
+            ),
+            (
+                {"output_1": "component_1.property_1", "output_2": "component_2.property_2"},
+                {"output_1": Output("component_1", "property_1"), "output_2": Output("component_2", "property_2")},
             ),
         ],
     )
@@ -166,8 +181,19 @@ class TestLegacyActionOutputs:
         [
             [""],
             ["component"],
+            ["component."],
+            [".property"],
+            ["component..property"],
             ["component_property"],
             ["component.property.property"],
+            {"output_1": ""},
+            {"output_1": "component"},
+            {"output_1": "component."},
+            {"output_1": ".property"},
+            {"output_1": "component..property"},
+            {"output_1": "component_property"},
+            {"output_1": "component.property.property"},
+            {"output_1": "component.property", "output_2": ""},
         ],
     )
     def test_outputs_invalid(self, outputs):
@@ -271,6 +297,26 @@ class TestActionInputs:
         action = Action(function=action_function(**inputs))
         assert action._transformed_inputs == expected_transformed_inputs
 
+    @pytest.mark.parametrize(
+        "input",
+        [
+            ["component.property"],
+            1,
+            None,
+            "",
+            "component",
+            "component.",
+            ".property",
+            "component..property",
+            "component_property",
+            "component.property.property",
+        ],
+    )
+    @pytest.mark.xfail(reason="Validation will only be performed once legacy actions are removed")
+    def test_runtime_inputs_invalid(self, input):
+        with pytest.raises(ValidationError):
+            Action(function=action_with_one_arg(input))._transformed_inputs
+
 
 class TestBuiltinRuntimeArgs:
     """Test the actual values of the runtime args are correct in a real scenario."""
@@ -290,6 +336,15 @@ class TestActionOutputs:
                 ["component_1.property_1", "component_2.property_2"],
                 [Output("component_1", "property_1"), Output("component_2", "property_2")],
             ),
+            ({}, {}),
+            (
+                {"output_1": "component.property"},
+                {"output_1": Output("component", "property")},
+            ),
+            (
+                {"output_1": "component_1.property_1", "output_2": "component_2.property_2"},
+                {"output_1": Output("component_1", "property_1"), "output_2": Output("component_2", "property_2")},
+            ),
         ],
     )
     def test_outputs_valid(self, outputs, expected_transformed_outputs):
@@ -301,14 +356,29 @@ class TestActionOutputs:
     @pytest.mark.parametrize(
         "outputs",
         [
+            "component.property",
+            1,
+            None,
             [""],
             ["component"],
+            ["component."],
+            [".property"],
+            ["component..property"],
             ["component_property"],
             ["component.property.property"],
+            {"output_1": ""},
+            {"output_1": "component"},
+            {"output_1": "component."},
+            {"output_1": ".property"},
+            {"output_1": "component..property"},
+            {"output_1": "component_property"},
+            {"output_1": "component.property.property"},
+            {"output_1": "component.property", "output_2": ""},
+            {1: "component.property"},
         ],
     )
     def test_outputs_invalid(self, outputs):
-        with pytest.raises(ValidationError, match="String should match pattern"):
+        with pytest.raises(ValidationError):
             Action(function=action_with_no_args(), outputs=outputs)