+++ vizro-core/tests/unit/vizro/models/_components/form/test_radio_items.py
@@ -178,7 +178,7 @@ class TestRadioItemsBuild:
             description=Tooltip(text="Test description", icon="info", id="info"),
         ).build()
 
-        description = [
+        expected_description = [
             html.Span("info", id="info-icon", className="material-symbols-outlined tooltip-icon"),
             dbc.Tooltip(
                 children=dcc.Markdown("Test description", className="card-text"),
@@ -190,7 +190,7 @@ class TestRadioItemsBuild:
 
         expected_radio_items = html.Fieldset(
             [
-                html.Legend(["Title", *description], className="form-label"),
+                html.Legend(["Title", *expected_description], className="form-label"),
                 dbc.RadioItems(
                     id="radio_items",
                     options=["A", "B", "C"],